+++ CHANGELOG.md
@@ -87,6 +87,7 @@ These PRs are not directly user-facing, but improve the development experience. 
 - **API**: Concepts now include the components they use in their doc comments.
 - **API**: `#[main]`-attributed functions no longer have to be `async` or return a `Result`.
 - **API**: `#[main]`-attributed functions, `on`, `once`, `Query::bind` and `run_async` can now return a `Result` or nothing.
+- **API**: Added `create_revolute_join` to the physics api.
 
 ### Fixed